Moved multi click safe views to one package
