+++ collect_app/src/main/res/layout/hierarchy_layout.xml
@@ -33,21 +33,21 @@ the License.
         android:layout_alignParentBottom="true"
         android:orientation="horizontal">
 
-        <org.odk.collect.android.views.MultiClickSafeButton
+        <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
             android:id="@+id/jumpBeginningButton"
             style="@style/Widget.Collect.Button.BottomOption"
             android:layout_marginLeft="@dimen/margin_extra_small"
             android:layout_marginRight="@dimen/margin_extra_small"
             android:text="@string/jump_to_beginning" />
 
-        <org.odk.collect.android.views.MultiClickSafeButton
+        <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
             android:id="@+id/jumpEndButton"
             style="@style/Widget.Collect.Button.BottomOption"
             android:layout_marginLeft="@dimen/margin_extra_small"
             android:layout_marginRight="@dimen/margin_extra_small"
             android:text="@string/jump_to_end" />
 
-        <org.odk.collect.android.views.MultiClickSafeButton
+        <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
             android:id="@+id/exitButton"
             style="@style/Widget.Collect.Button.BottomOption"
             android:layout_marginLeft="@dimen/margin_extra_small"
