+++ collect_app/src/main/res/layout/main_menu.xml
@@ -41,7 +41,7 @@
                 android:paddingRight="@dimen/margin_standard"
                 android:paddingBottom="@dimen/margin_standard">
 
-                <org.odk.collect.android.views.MultiClickSafeButton
+                <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
                     android:id="@+id/enter_data"
                     style="@style/Widget.Collect.Button.FullWidth"
                     android:layout_width="match_parent"
@@ -49,7 +49,7 @@
                     android:layout_marginTop="@dimen/margin_extra_small"
                     tools:text="@string/enter_data" />
 
-                <org.odk.collect.android.views.MultiClickSafeButton
+                <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
                     android:id="@+id/review_data"
                     style="@style/Widget.Collect.Button.FullWidth"
                     android:layout_width="match_parent"
@@ -57,7 +57,7 @@
                     android:layout_marginTop="@dimen/margin_extra_small"
                     tools:text="@string/review_data" />
 
-                <org.odk.collect.android.views.MultiClickSafeButton
+                <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
                     android:id="@+id/send_data"
                     style="@style/Widget.Collect.Button.FullWidth"
                     android:layout_width="match_parent"
@@ -65,7 +65,7 @@
                     android:layout_marginTop="@dimen/margin_extra_small"
                     tools:text="@string/send_data" />
 
-                <org.odk.collect.android.views.MultiClickSafeButton
+                <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
                     android:id="@+id/view_sent_forms"
                     style="@style/Widget.Collect.Button.FullWidth"
                     android:layout_width="match_parent"
@@ -73,7 +73,7 @@
                     android:layout_marginTop="@dimen/margin_extra_small"
                     android:text="@string/view_sent_forms" />
 
-                <org.odk.collect.android.views.MultiClickSafeButton
+                <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
                     android:id="@+id/get_forms"
                     style="@style/Widget.Collect.Button.FullWidth"
                     android:layout_width="match_parent"
@@ -81,7 +81,7 @@
                     android:layout_marginTop="@dimen/margin_extra_small"
                     tools:text="@string/get_forms" />
 
-                <org.odk.collect.android.views.MultiClickSafeButton
+                <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
                     android:id="@+id/manage_forms"
                     style="@style/Widget.Collect.Button.FullWidth"
                     android:layout_width="match_parent"
