+++ collect_app/src/main/res/layout/osm_widget_answer.xml
@@ -5,7 +5,7 @@
     android:paddingTop="@dimen/margin_small"
     android:paddingHorizontal="@dimen/margin_standard">
 
-    <org.odk.collect.android.views.MultiClickSafeButton
+    <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
         android:id="@+id/launch_open_map_kit_button"
         style="@style/Widget.Collect.Button.WidgetAnswer"
         android:text="@string/capture_osm"
