+++ collect_app/src/main/res/layout/url_widget_answer.xml
@@ -5,7 +5,7 @@
     android:paddingTop="@dimen/margin_small"
     android:paddingHorizontal="@dimen/margin_standard">
 
-    <org.odk.collect.android.views.MultiClickSafeButton
+    <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
         android:id="@+id/url_button"
         style="@style/Widget.Collect.Button.WidgetAnswer"
         android:text="@string/open_url" />
